refactor: pointerlock
