+++ packages/utilities/dom-event/src/index.ts
@@ -4,7 +4,7 @@ export * from "./get-event-step"
 export * from "./get-native-event"
 export * from "./add-dom-event"
 export * from "./point-value"
-export * from "./pointer-lock"
-export * from "./pointer-move"
+export * from "./track-pointer-lock"
+export * from "./track-pointer-move"
 export * from "./fire-event"
 export type { EventKeyMap } from "./types"
