+++ packages/utilities/dom-event/src/track-pointer-move.ts
